@@ -12,25 +12,28 @@
   "dependencies": {
     "@headlessui/react": "^1.7.7",
     "@mapbox/rehype-prism": "^0.7.0",
-    "@next/mdx": "^13.1.1",
-    "@supabase/auth-helpers-nextjs": "^0.6.0",
-    "@supabase/auth-helpers-react": "^0.3.1",
+    "@mdx-js/loader": "^2.3.0",
+    "@mdx-js/react": "^2.3.0",
+    "@next/mdx": "^13.4.10",
+    "@supabase/auth-helpers-nextjs": "^0.7.3",
+    "@supabase/supabase-js": "^2.27.0",
     "@tailwindcss/typography": "^0.5.8",
-    "@types/mdx": "^2.0.3",
+    "@types/mdx": "^2.0.5",
     "@types/node": "^18.11.18",
     "@types/nprogress": "^0.2.0",
     "@types/react": "18.0.26",
     "@types/react-dom": "18.0.10",
     "autoprefixer": "^10.4.13",
     "clsx": "^1.2.1",
+    "encoding": "^0.1.13",
     "eslint": "8.31.0",
     "eslint-config-next": "^13.3.0",
     "fast-glob": "^3.2.12",
     "feed": "^4.2.2",
     "focus-visible": "^5.2.0",
     "motion": "^10.15.5",
-    "next": "^13.3.0",
-    "node-fetch": "^3.3.0",
+    "next": "^13.4.10",
+    "next-themes": "^0.2.1",
     "nprogress": "^0.2.0",
     "postcss": "^8.4.21",
     "postcss-focus-visible": "^7.1.0",
@@ -38,13 +41,11 @@
     "react-dom": "^18.2.0",
     "remark-gfm": "^3.0.1",
     "sharp": "^0.31.3",
+    "supabase": "^1.50.2",
     "swr": "^2.1.2",
     "tailwind-merge": "^1.9.0",
     "tailwindcss": "^3.3.0",
     "ts-node": "^10.9.1",
     "typescript": "4.9.4"
-  },
-  "devDependencies": {
-    "supabase": "^1.50.2"
   }
 }
