@@ -1,47 +1,11 @@
-import {useRouter} from 'next/router'
+'use client'
+
 import {useEffect, useRef} from 'react'
+import {usePathname} from 'next/navigation'
 import {Container} from './Container'
-import {MobileNavigation, DesktopNavigation} from '@/components/ui/Navigation'
+import {ThemeButton} from '@/components/ui/ThemeButton'
+import {DesktopNavigation, MobileNavigation} from '@/components/ui/Navigation'
 import {Avatar, AvatarContainer} from '@/components/ui/Avatar'
-import {MoonIcon} from '@/components/icons/MoonIcon'
-import {SunIcon} from '@/components/icons/SunIcon'
-
-function ModeToggle() {
-    function disableTransitionsTemporarily() {
-        document.documentElement.classList.add('[&_*]:!transition-none')
-        window.setTimeout(() => {
-            document.documentElement.classList.remove('[&_*]:!transition-none')
-        }, 0)
-    }
-
-    function toggleMode() {
-        disableTransitionsTemporarily()
-
-        let darkModeMediaQuery = window.matchMedia('(prefers-color-scheme: dark)')
-        let isSystemDarkMode = darkModeMediaQuery.matches
-        let isDarkMode = document.documentElement.classList.toggle('dark')
-
-        if (isDarkMode === isSystemDarkMode) {
-            delete window.localStorage.isDarkMode
-        } else {
-            window.localStorage.isDarkMode = isDarkMode
-        }
-    }
-
-    return (
-        <button
-            type="button"
-            aria-label="Toggle dark mode"
-            className="group rounded-full bg-white/90 px-3 py-2 shadow-lg shadow-zinc-800/5 ring-1 ring-zinc-900/5 backdrop-blur transition dark:bg-zinc-800/90 dark:ring-white/10 dark:hover:ring-white/20"
-            onClick={toggleMode}
-        >
-            <SunIcon
-                className="h-6 w-6 fill-zinc-100 stroke-zinc-500 transition group-hover:fill-zinc-200 group-hover:stroke-zinc-700 dark:hidden [@media(prefers-color-scheme:dark)]:fill-indigo-50 [@media(prefers-color-scheme:dark)]:stroke-indigo-500 [@media(prefers-color-scheme:dark)]:group-hover:fill-indigo-50 [@media(prefers-color-scheme:dark)]:group-hover:stroke-indigo-600"/>
-            <MoonIcon
-                className="hidden h-6 w-6 fill-zinc-700 stroke-zinc-500 transition dark:block [@media(prefers-color-scheme:dark)]:group-hover:stroke-zinc-400 [@media_not_(prefers-color-scheme:dark)]:fill-indigo-400/10 [@media_not_(prefers-color-scheme:dark)]:stroke-indigo-500"/>
-        </button>
-    )
-}
 
 function clamp(num: number, a: number, b: number) {
     let min = Math.min(a, b)
@@ -50,8 +14,8 @@ function clamp(num: number, a: number, b: number) {
 }
 
 export function Header() {
-    const router = useRouter()
-    const isHomePage = router.pathname === '/'
+    const pathname = usePathname()
+    const isHomePage = pathname === '/'
 
     const headerRef = useRef<HTMLDivElement>(null)
     const avatarRef = useRef<HTMLImageElement>(null)
@@ -227,7 +191,7 @@ export function Header() {
                             </div>
                             <div className="flex justify-end md:flex-1">
                                 <div className="pointer-events-auto">
-                                    <ModeToggle/>
+                                    <ThemeButton/>
                                 </div>
                             </div>
                         </div>
