@@ -18,8 +18,8 @@ export async function generateRssFeed() {
         author,
         id: siteUrl!,
         link: siteUrl,
-        image: `${siteUrl}/static/icons/favicon.ico`,
-        favicon: `${siteUrl}/static/icons/favicon.ico`,
+        image: `${siteUrl}/favicon.ico`,
+        favicon: `${siteUrl}/favicon.ico`,
         copyright: `All rights reserved ${new Date().getFullYear()}`,
         feedLinks: {
             rss2: `${siteUrl}/rss/feed.xml`,
